// client/src/pages/client-detail/types/measurement.ts
export interface Measurement {
  id: number;
  date: string;
  weight: string;
  height: string;
  waistCircumference?: string;
  hipCircumference?: string;
  bodyFatPercentage?: string;
  activityLevel: string;
  notes?: string;
  bmi: string;
  basalMetabolicRate: number;
  totalDailyEnergyExpenditure: number;
  // Micro-nutrients
  vitaminA?: string;
  vitaminC?: string;
  vitaminD?: string;
  vitaminE?: string;
  vitaminK?: string;
  thiamin?: string;
  riboflavin?: string;
  niacin?: string;
  vitaminB6?: string;
  folate?: string;
  vitaminB12?: string;
  biotin?: string;
  pantothenicAcid?: string;
  calcium?: string;
  iron?: string;
  magnesium?: string;
  phosphorus?: string;
  zinc?: string;
  potassium?: string;
  sodium?: string;
  copper?: string;
  manganese?: string;
  selenium?: string;
  chromium?: string;
  molybdenum?: string;
  iodine?: string;
} 